backend/pyproject.toml
[project]
name = "querywise"
version = "0.1.0"
description = "QueryWise — Natural language to SQL with a semantic metadata layer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pgvector>=0.3",
    "httpx>=0.27",
    "sse-starlette>=2.0",
    "sqlparse>=0.5",
    "cryptography>=43.0",
    "PyJWT>=2.8",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0",
    "email-validator>=2.0",
    "loguru>=0.7",
    "platformdirs>=4.0",
    "tenacity>=8.0",
    "prometheus-fastapi-instrumentator>=7.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.40",
    "openai>=1.50",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langsmith>=0.1",
]
sqlserver = [
    "aioodbc>=0.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.setuptools.packages.find]
include=["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
# graph.py contains an ASCII-art topology diagram in its module docstring
"app/llm/graph/graph.py" = ["E501"]
# Prompts are natural language, not Python code
"app/llm/prompts/*.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end integration tests (require a running backend and real database connection)",
]
